backend/requirements.txt
flask
flask-cors
groq
python-dotenv
reportlab
gunicorn
